New Class: GCode Commands
This class will manage all single GCode Commands
in ASCII format.
Its a Enum list with all commands in string and command form.
The string form is to explain what the command do.
The command form is the command pre-build for send
to the serial port.
Implemented the base G commands based on RepRap GCode
Wiki.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>