Finished first implementation of a plugin system.
Ok, the code currently is a bit of a mess - the AtCore clearly
lacks a good architecture and we must stop what we are doing now
to think a bit about the architecture that we want to use.

Currently too much of setup is done within the test application,
this should be used inside of the AtCore - a init method perhaps.

I don't know what should be inside the firmwares, so one with
more knowledge of printing functions should add what's missing there

Right now there's a Protocol Layer / Firmware Plugin that are
side by side in the d-pointer, and after a brief talk with Patrick
that's ok, but the plugin should translate what will be send to the
ProtocolLayer and vice versa - that's what I don't have any idea
on how to do.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@gmail.com>