Dock based interface, Create a status bar.
Summary:
This the is the general design I have in mind for atelier adapted to the test client .
The gui now uses serveral QDockWidgets to hold the various controls and also contains a status bar with some basic info like atcore state and the printLayout has also been moved to the status bar. The gui now also contains one menu view , where you can toggle the docks visibility (incase you close one you want to open again) .

UPDATE:
ToDo before merging.

  - [ ] Fix View Menu (see test plan update)
  - [X] Remove the settings menu . We dont want the user to hide the status bar and we don't provide any action for ShortCut or Toolbars
  - [ ] Action to show /hide the dock handles

Test Plan:
The only issues with this gui is that when i connect the docks visibilityChanged to the actions setChecked and the actions toggled to the docks setVisibility you lose the ability to tabify them since when they are not the active tabified dock they close because of the toggle of the action. Something is wrong in how im doing this menu and that needs to be fixed before this can be merged. To see the behavior you need to uncomment the connects for the docks signal in testclient/mainwindow.cpp  .

UPDATE:
This is better but not perfect with KXMLGui. Now the problem is that when tabified the items within the view menu are only check if the widget is the selected tab. This is much better then them disapearing but we still need to address this problem before merging this.

Reviewers: laysrodrigues, patrickelectric, tcanabrava

Differential Revision: https://phabricator.kde.org/D5529